requirements.txt
mongo
python-dotenv
gunicorn
python-telegram-bot==21.4
openai
